feat: 实现用户头像上传功能并重构缩略图存储
添加用户头像上传功能，支持通过API上传并转换为AVIF格式
重构缩略图存储系统，将原有缩略图路径替换为统一资源存储
新增assets表用于管理所有资源文件，包括用户头像和归档封面
移除旧的THUMBNAIL_PATH配置，迁移现有缩略图到新存储系统
添加多语言支持，完善前端头像上传和显示功能

+++ frontend/src/components/ui/slider.tsx
@@ -15,7 +15,7 @@ interface SliderProps extends Omit<React.InputHTMLAttributes<HTMLInputElement>, 
 const Slider = React.forwardRef<HTMLInputElement, SliderProps>(
   ({ className, value = [0], onValueChange, max = 100, min = 0, step = 1, ...props }, ref) => {
     const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-      const newValue = [parseInt(e.target.value)];
+      const newValue = [parseFloat(e.target.value)];
       onValueChange?.(newValue);
     };
